% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANECO-package.R
\docType{package}
\name{ANECO-package}
\alias{ANECO-package}
\title{ANECO: Tools for the analysis of ecoacoustic indices}
\description{
ANECO is an R package for ecoacoustic analysis that streamlines the calculation of nearly 30 acoustic indices, including underexplored metrics. It offers tools for sound pressure calibration and automated detection of rain dominated recordings. Designed for ease of use, ANECO supports efficient workflows and encourages multi-index approaches for robust ecological interpretation.
}
\keyword{internal}
