% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indices_codes}
\alias{indices_codes}
\title{Index codes available in ANECO}
\format{
Character vector.
}
\usage{
data(indices_codes)
}
\description{
Character vector of valid index codes that can be passed to
`acoustic_indices()` through the `sel.ind` argument.
}
\keyword{datasets}
