% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R, R/dataset.R
\docType{data}
\name{indices_data}
\alias{indices_data}
\title{Example acoustic indices dataset}
\format{
A data frame where rows represent minutes and
columns represent acoustic indices and metadata.

An object of class \code{data.frame} with 7770 rows and 69 columns.
}
\source{
Example dataset included in ANECO
}
\usage{
indices_data

indices_data
}
\description{
Acoustic indices calculated using acoustic_indices()

A dataset of acoustic indices resulting from the function \code{\link[ANECO]{acoustic_indices}}, with the parameter \code{"noise.ind = TRUE"}
}
\keyword{datasets}
