% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R, R/dataset.R
\docType{data}
\name{labeled_data}
\alias{labeled_data}
\title{Example labeled data for sound classification}
\format{
A data frame with 4 columns:
\describe{
  \item{File}{Audio file name}
  \item{Category}{Label (Clean, Rain, Cicada)}
  \item{Start}{Start minute}
  \item{End}{End minute}
}

A dataframe with the columns File, Category, Start, and End. Where:
\describe{
  \item{File}{Full name of the audio file, including the .wav extension. For example: FT02_20190615_110000.wav}
  \item{Category}{Category of the audio segment. Three categories are available: Clean, Rain, Cicada}
  \item{Start}{Start time of the event in the analyzed category, with a resolution of 0.5 seconds}
  \item{End}{End time of the event in the analyzed category, with a resolution of 0.5 seconds}
}
}
\source{
Example dataset included in ANECO
}
\usage{
labeled_data

labeled_data
}
\description{
A dataset containing manually labeled audio segments used
to train the random forest classifier in ANECO.

A dataset indicating the periods with occurrences of rain and cicada sounds.
}
\keyword{datasets}
